+++ backend/favorite_things/helpers.py
@@ -1,0 +1,15 @@
+def split_choices_int(choices):
+    assert(isinstance(choices, list))
+    include = []
+    exclude = []
+    for t in choices:
+        try:
+            t = int(t)
+            if t < 0:
+                exclude.append(-t)
+            else:
+                include.append(t)
+        except:
+            pass
+
+    return include, exclude
