+++ frontend/src/index.css
@@ -17,7 +17,7 @@ body{
 body{
     font-family: 'Open Sans', sans-serif;
     background-color: #f7f7f7;
-    background-image: url("./assets/trackit-min.jpg");
+    background-image: url("https://res.cloudinary.com/ladiesintech/image/upload/v1568302244/favorite/trackit-min.jpg");
     background-repeat: no-repeat;
     background-position: center;
     background-size: cover;
